moved validations to validators and refactor code
