+++ src/main/java/com/bjit/mailservice/services/impl/SendGridMailService.java
@@ -1,7 +1,7 @@
 package com.bjit.mailservice.services.impl;
 
 import com.bjit.mailservice.services.MailService;
-import com.bjit.mailservice.services.MailValidation;
+import com.bjit.mailservice.validators.MailValidation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ClassPathResource;
