--- src/main/java/com/bjit/mailservice/utils/ValidAttachment.java
@@ -1,6 +1,0 @@
-package com.bjit.mailservice.utils;
-
-import jakarta.validation.groups.Default;
-
-public interface ValidAttachment extends Default {
-}
